requirements/development.txt
-r server.txt
django-debug-toolbar